backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
joblib
supabase==1.2.0
python-dotenv
